Fix csv to geojson converter for traficom datasets
In order of appearance in the diff:

Without having the datasets from vayla the propsToCopy name changes are
educated guesses based on the values in those columns in the respective
datasets.

The SHAPE colunn is called GEOM in traficom.

The syvyysalue.csv from traficom exposed a bug in the rudimentary way
the SHAPE/GEOM column value is transformed into JSON. I can
t be arsed to write a proper converter now, so this dirty patch will
have to do for now.